requirements.txt
pandas 
numpy>=1.9.2 
flask 
scikit-learn
seaborn
datetime
typing
pickles
logging
streamlit
mysql-connector-python
-e .